requirements.txt
# Core dependencies
pymilvus>=2.3.0
openai>=1.0.0
python-dotenv>=1.0.0
diskcache>=5.6.0

# Optional dependencies for development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Type checking (optional)
mypy>=1.0.0

# Code formatting (optional)
black>=23.0.0
isort>=5.12.0
